Moving tests to proposals_for_votainteligente
