+++ popular_proposal/tests/proposal_search_tests.py
@@ -4,7 +4,6 @@ from popular_proposal.tests import ProposingCycleTestCaseBase
 from popular_proposal.search_indexes import ProposalIndex
 from popular_proposal.forms.form_texts import TOPIC_CHOICES
 from django.core.management import call_command
-from elections.models import Area
 from popular_proposal.forms.form_texts import TOPIC_CHOICES
 import haystack
 from haystack.query import SearchQuerySet
@@ -13,21 +12,17 @@ from haystack.query import SearchQuerySet
 class PopularProposalSearchIndexTestCase(ProposingCycleTestCaseBase):
     def setUp(self):
         super(PopularProposalSearchIndexTestCase, self).setUp()
-        self.algarrobo = Area.objects.get(id='algarrobo-5602')
         self.p1 = PopularProposal.objects.create(proposer=self.fiera,
-                                                 area=self.algarrobo,
                                                  data=self.data,
                                                  title=u'Que solucionen los problemas',
                                                  clasification=TOPIC_CHOICES[1][0]
                                                  )
         self.p2 = PopularProposal.objects.create(proposer=self.fiera,
-                                                 area=self.algarrobo,
                                                  data=self.data,
                                                  title=u'Bicicletas para todas',
                                                  clasification=TOPIC_CHOICES[2][0]
                                                  )
         self.p3 = PopularProposal.objects.create(proposer=self.fiera,
-                                                 area=self.algarrobo,
                                                  data=self.data,
                                                  title=u'Educación de caludad',
                                                  clasification=TOPIC_CHOICES[3][0]
@@ -58,7 +53,6 @@ class PopularProposalSearchIndexTestCase(ProposingCycleTestCaseBase):
                            'clasification': TOPIC_CHOICES[1][0],
                            'causes': u'Perros'}
         p1 = PopularProposal.objects.create(proposer=self.fiera,
-                                                 area=self.algarrobo,
                                                  data=data_with_texts,
                                                  title=u'Que solucionen los problemas',
                                                  clasification=TOPIC_CHOICES[1][0]
