+++ popular_proposal/tests/temporary_proposal_tests.py
@@ -1,4 +1,3 @@
-
 # coding=utf-8
 from popular_proposal.tests import ProposingCycleTestCaseBase
 from django.contrib.auth.models import User
@@ -18,7 +17,6 @@ class TemporaryDataForPromise(ProposingCycleTestCaseBase):
     def test_instanciate_one(self):
         temporary_data = ProposalTemporaryData.objects.create(proposer=self.fiera,
                                                               join_advocacy_url=u"http://whatsapp.com/mygroup",
-                                                              area=self.arica,
                                                               data=self.data)
         self.assertTrue(temporary_data)
         self.assertTrue(temporary_data.join_advocacy_url)
@@ -39,7 +37,6 @@ class TemporaryDataForPromise(ProposingCycleTestCaseBase):
 
     def test_send_temporary_data_new_mail(self):
         temporary_data = ProposalTemporaryData.objects.create(proposer=self.fiera,
-                                                              area=self.arica,
                                                               data=self.data)
         temporary_data.notify_new()
         self.assertEquals(len(mail.outbox), 1)
@@ -49,14 +46,11 @@ class TemporaryDataForPromise(ProposingCycleTestCaseBase):
 
     def test_needing_moderation_proposals(self):
         td_waiting_for_moderation = ProposalTemporaryData.objects.create(proposer=self.fiera,
-                                                                         area=self.arica,
                                                                          data=self.data)
         td_waiting_for_moderation2 = ProposalTemporaryData.objects.create(proposer=self.fiera,
-                                                                          area=self.arica,
                                                                           data=self.data)
         needs_citizen_action = ProposalTemporaryData.objects.create(proposer=self.fiera,
                                                                     status=ProposalTemporaryData.Statuses.InTheirSide,
-                                                                    area=self.arica,
                                                                     data=self.data)
         self.assertIn(td_waiting_for_moderation, ProposalTemporaryData.needing_moderation.all())
         self.assertIn(td_waiting_for_moderation2, ProposalTemporaryData.needing_moderation.all())
@@ -64,7 +58,6 @@ class TemporaryDataForPromise(ProposingCycleTestCaseBase):
 
     def test_rejecting_a_proposal(self):
         temporary_data = ProposalTemporaryData.objects.create(proposer=self.fiera,
-                                                              area=self.arica,
                                                               data=self.data)
         original_amount_of_mails = len(mail.outbox)
         temporary_data.reject('es muy mala la cosa')
@@ -86,37 +79,3 @@ class TemporaryDataForPromise(ProposingCycleTestCaseBase):
         # template_subject.render(context)
         self.assertTrue(the_mail.body)
         self.assertTrue(the_mail.subject)
-
-
-class ProposingViewTestCase(ProposingCycleTestCaseBase):
-    def setUp(self):
-        super(ProposingViewTestCase, self).setUp()
-        self.feli = User.objects.get(username='feli')
-        self.feli.set_password('alvarez')
-        self.feli.save()
-
-    def test_get_proposing_view(self):
-        url = reverse('popular_proposals:propose', kwargs={'slug': self.arica.id})
-        #need to be loggedin
-        response = self.client.get(url)
-        self.assertEquals(response.status_code, 302)
-
-        self.client.login(username=self.feli,
-                          password='alvarez')
-        response = self.client.get(url)
-        self.assertEquals(response.status_code, 200)
-        form = response.context['form']
-        self.assertEquals(self.arica, response.context['area'])
-        self.assertIsInstance(form, ProposalForm)
-
-    def test_post_proposing_view(self):
-        url = reverse('popular_proposals:propose', kwargs={'slug': self.arica.id})
-
-        self.client.login(username=self.feli,
-                          password='alvarez')
-        self.assertFalse(ProposalTemporaryData.objects.all())
-        response = self.client.post(url, data=self.data, follow=True)
-        self.assertEquals(response.status_code, 200)
-        self.assertTemplateUsed('popular_proposal/thanks.html')
-        temporary_data = ProposalTemporaryData.objects.get()
-        self.assertTrue(temporary_data)
