+++ proposals_for_votainteligente/tests/temporary_proposal_tests.py
@@ -1,0 +1,44 @@
+# coding=utf-8
+from popular_proposal.tests import ProposingCycleTestCaseBase
+from django.contrib.auth.models import User
+from popular_proposal.models import ProposalTemporaryData
+from popular_proposal.forms import ProposalForm
+from django.core.urlresolvers import reverse
+from django.core import mail
+from django.contrib.staticfiles.templatetags.staticfiles import static
+from django.template.loader import get_template
+from django.contrib.sites.models import Site
+
+
+class ProposingViewTestCase(ProposingCycleTestCaseBase):
+    def setUp(self):
+        super(ProposingViewTestCase, self).setUp()
+        self.feli = User.objects.get(username='feli')
+        self.feli.set_password('alvarez')
+        self.feli.save()
+
+    def test_get_proposing_view(self):
+        url = reverse('popular_proposals:propose', kwargs={'slug': self.arica.id})
+        #need to be loggedin
+        response = self.client.get(url)
+        self.assertEquals(response.status_code, 302)
+
+        self.client.login(username=self.feli,
+                          password='alvarez')
+        response = self.client.get(url)
+        self.assertEquals(response.status_code, 200)
+        form = response.context['form']
+        self.assertEquals(self.arica, response.context['area'])
+        self.assertIsInstance(form, ProposalForm)
+
+    def test_post_proposing_view(self):
+        url = reverse('popular_proposals:propose', kwargs={'slug': self.arica.id})
+
+        self.client.login(username=self.feli,
+                          password='alvarez')
+        self.assertFalse(ProposalTemporaryData.objects.all())
+        response = self.client.post(url, data=self.data, follow=True)
+        self.assertEquals(response.status_code, 200)
+        self.assertTemplateUsed('popular_proposal/thanks.html')
+        temporary_data = ProposalTemporaryData.objects.get()
+        self.assertTrue(temporary_data)
